requirements.txt
transformers>=4.35
torch>=2.0
datasets
sentence-transformers
faiss-cpu
rdflib
PyYAML
python-dotenv
requests
beautifulsoup4
lxml
numpy
pandas
tqdm
biopython
accelerate
scikit-learn
matplotlib
seaborn
jupyter
ipywidgets
